Progress in video now works
